+++ cogeo_mosaic/backends/__init__.py
@@ -6,6 +6,7 @@ from urllib.parse import urlparse
 from cogeo_mosaic.backends.base import BaseBackend
 from cogeo_mosaic.backends.dynamodb import DynamoDBBackend
 from cogeo_mosaic.backends.file import FileBackend
+from cogeo_mosaic.backends.memory import InMemoryBackend
 from cogeo_mosaic.backends.s3 import S3Backend
 from cogeo_mosaic.backends.sqlite import SQLiteBackend
 from cogeo_mosaic.backends.stac import STACBackend
@@ -16,8 +17,11 @@ def MosaicBackend(url: str, *args: Any, **kwargs: Any) -> BaseBackend:
     """Select mosaic backend for url."""
     parsed = urlparse(url)
 
+    if not url or url == ":memory:":
+        return InMemoryBackend(*args, **kwargs)
+
     # `stac+https//{hostname}/{path}`
-    if parsed.scheme and parsed.scheme.startswith("stac+"):
+    elif parsed.scheme and parsed.scheme.startswith("stac+"):
         url = url.replace("stac+", "")
         return STACBackend(url, *args, **kwargs)