+++ tests/test_backends.py
@@ -18,6 +18,7 @@ from requests.exceptions import HTTPError, RequestException
 from cogeo_mosaic.backends import MosaicBackend
 from cogeo_mosaic.backends.dynamodb import DynamoDBBackend
 from cogeo_mosaic.backends.file import FileBackend
+from cogeo_mosaic.backends.memory import InMemoryBackend
 from cogeo_mosaic.backends.s3 import S3Backend
 from cogeo_mosaic.backends.sqlite import SQLiteBackend
 from cogeo_mosaic.backends.stac import STACBackend
@@ -639,16 +640,22 @@ def test_mosaic_crud_error(mosaic_path):
             ...
 
 
-def test_BaseReader():
-    """Test BaseReader heritance methods."""
+def test_InMemoryReader():
+    """Test InMemoryBackend."""
     assets = [asset1, asset2]
     mosaicdef = MosaicJSON.from_urls(assets, quiet=False)
 
-    # add some offset to the center to make
-    # sure BaseBackend forward center from the mosaic definition
-    mosaicdef.center = [x + 1 for x in mosaicdef.center]
+    with MosaicBackend(":memory:", mosaic_def=mosaicdef) as mosaic:
+        assert isinstance(mosaic, InMemoryBackend)
+        assert mosaic.path == ":memory:"
+        mosaic.write()
+        mosaic._read()
 
     with MosaicBackend(None, mosaic_def=mosaicdef) as mosaic:
+        assert isinstance(mosaic, InMemoryBackend)
+        assert mosaic.path == ":memory:"
+
+    with InMemoryBackend(mosaic_def=mosaicdef) as mosaic:
         (t, _), assets_used = mosaic.tile(150, 182, 9)
         assert t.shape
 
@@ -704,6 +711,17 @@ def test_BaseReader():
         ]
 
         assert mosaic.spatial_info
+
+    mosaic_oneasset = MosaicJSON.from_urls([asset1], quiet=True)
+    with InMemoryBackend(mosaic_def=mosaic_oneasset) as mosaic:
+        assert isinstance(mosaic, InMemoryBackend)
+        assert len(mosaic.get_assets(150, 182, 9)) == 1
+        features = get_footprints([asset2], quiet=True)
+        mosaic.update(features)
+        assets = mosaic.get_assets(150, 182, 9)
+        assert len(assets) == 2
+        assert assets[0] == asset2
+        assert assets[1] == asset1
 
 
 def test_sqlite_backend():